feat: catch likely usage mistake with command as first argument
If a user runs something like `run0edit vim /path/to/file`, this is
likely a usage error and they intend to edit `/path/to/file` using vim,
*not* edit a file at `./vim` and then edit a second file.

This adds detection for this scenario by warning the user and prompting
for confirmation if there are multiple positional arguments, the first
of which does not contain a path separator and would resolve to an
executable if interpreted as a command. The warning is suppressed by the
`--no-prompt` option.