fix: prevent SELinux denial with certain custom editors (#9)
Previously, if the user chose an editor that didn't have the right
SELinux label (e.g. if it was in the user's home directory instead of a
directory like /usr/bin labeled for executables), the inner run0 command
to run the editor would fail with status 203. This works around the
issue by wrapping the editor command in `sh -c`.