fix: adjust sandbox to fix issue creating files in certain directories
It appears that `ReadWritePaths` will not successfully set a directory
as writable if it's already been directly set as read-only by
`ProtectHome=read-only`. This affects creating files in `/root` and
`/home`, but only on systems where these are not symlinks, and only
*directly* in these directories (not creating files in subdirectories).

To fix this, we replace `ProtectSystem=strict` and
`ProtectHome=read-only` with `ReadOnlyPaths=/`, which should provide
equally strong sandboxing but not incorrectly override the writable
directories set by `ReadWritePaths`.