chore!: drop support for Python 3.9
As of October 2025, Python 3.9 is end-of-life and no longer receiving
security updates. For that reason, run0edit is dropping Python 3.9
support and now has a minimum required Python version of 3.10.

This allows for a couple changes that improve code readability: the use
of `|` instead of `Union` for sum type annotations, and the use of
`match` blocks.