feat: Add basic validation to --background input
The argument to --background must be either the empty string or a
sequence of ASCII numbers separated by semicolons. Inputs not fitting
this format will be rejected with an error message. Note that this will
still accept inputs that don't correspond to valid ANSI formatting
codes, but it should at least prevent the terminal output from being
badly broken.