feat: Add --background option to set background color (#38)
* feat: Add --background option to set background color

Add optional `--background` argument to set or disable the background
color. This behaves the same as the `--background` option of `run0`, and
is implemented by passing through the option to the inner call to
`run0`.

Also update tests, changelog, and readme for the new option.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>

* feat: Add basic validation to --background input

The argument to --background must be either the empty string or a
sequence of ASCII numbers separated by semicolons. Inputs not fitting
this format will be rejected with an error message. Note that this will
still accept inputs that don't correspond to valid ANSI formatting
codes, but it should at least prevent the terminal output from being
badly broken.

---------

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>