build: add editor config file to RPM spec (#52)
This ensures that `/etc/run0edit/editor.conf` will be created (as an
empty file, which is ignored by `run0edit`) with the appropriate
permissions on install if not already present. This avoids the confusing
user experience of either the conf file or its parent directory being
created with the wrong permissions by default if the umask is set to
something stricter than 022.